refine src and header
